[#63] - Create AuthenticationTestsTrait.php
